refactor : change names of some variables and improving code
