+++ src/app.ts
@@ -27,7 +27,7 @@ export class ExpressApp {
 
   public start(port: number | string) {
     return this.expressApp.listen(port, () => {
-      logging.info(`Sever is running on port ${port}!`);
+      logging.info(`Server is running on port ${port}!`);
     });
   }
