+++ src/application/usecases/activity/create-activity-use-case.ts
@@ -11,7 +11,7 @@ export interface CreateActivity {
 export class CreateActivityUseCase implements CreateActivity {
   constructor(
     readonly activityRepository: IActivityRepository,
-    readonly logging: Logging
+    readonly logging: Logging,
   ) {}
   async invoke(data: Activity): Promise<Activity> {
     try {
@@ -19,11 +19,11 @@ export class CreateActivityUseCase implements CreateActivity {
       return await this.activityRepository.create(data);
     } catch (error) {
       this.logging.error(
-        `Some error has been ocurred trying create a new activity ${error}`
+        `Some error has been ocurred trying create a new activity ${error}`,
       );
       throw new AppError(
         "Internal server error",
-        HttpStatusCode.InternalServerError
+        HttpStatusCode.InternalServerError,
       );
     }
   }
