+++ src/application/usecases/activity/list-activity-use-case.ts
@@ -11,18 +11,18 @@ export interface ListActivity {
 export class ListActivityUseCase implements ListActivity {
   constructor(
     readonly activityRepository: IActivityRepository,
-    readonly logging: Logging
+    readonly logging: Logging,
   ) {}
   async invoke(user_id: string): Promise<Activity[] | null> {
     try {
       return await this.activityRepository.findMany(user_id);
     } catch (error) {
       this.logging.error(
-        `Some error has been ocurred trying  get a list of activities ${error}`
+        `Some error has been ocurred trying  get a list of activities ${error}`,
       );
       throw new AppError(
         "Internal server error",
-        HttpStatusCode.InternalServerError
+        HttpStatusCode.InternalServerError,
       );
     }
   }
