+++ src/infraestructure/factories/controllers/user/make-add-controller.ts
@@ -1,5 +1,5 @@
 import { Controller } from "../../../../domain/controller";
-import { CreateUserController } from "../../../../presentation/controllers/users/CreateUserController";
+import { CreateUserController } from "../../../../presentation/controllers/users/create-user-controller";
 import { makeDbAddUserUseCase } from "../../usecases/user/make-db-add-user";
 
 export const makeAddUserController = (): Controller => {
