+++ src/presentation/routes/authentication/auth-routes.ts
@@ -1,13 +1,5 @@
-import { Request, Router } from "express";
-import { authController } from "../../../infraestructure/config/LoginDependencie";
+import { Router } from "express";
 
 export const authRoutes = Router();
 
-authRoutes.post("/", async (req: Request, res) => {
-  const httpResponse = await authController.handle(req);
-  res.status(httpResponse.statusCode).json({
-    message: httpResponse.msg,
-    code: httpResponse.statusCode,
-    body: httpResponse.body,
-  });
-});
+authRoutes.post("/");
