+++ src/presentation/routes/users/user-routes.ts
@@ -1,13 +1,7 @@
-import { Request, Router } from "express";
-import { createUserController } from "../../../infraestructure/config/UsersDependencie";
+import { Router } from "express";
+import { adaptRoute } from "../../../adapters/router-adapter";
+import { makeAddUserController } from "../../../infraestructure/factories/controllers/user/make-add-controller";
 
 export const userRoutes = Router();
 
-userRoutes.post("/", async (req: Request, res) => {
-  const httpResponse = await createUserController.handle(req);
-  res.status(httpResponse.statusCode).json({
-    message: httpResponse.msg,
-    code: httpResponse.statusCode,
-    body: httpResponse.body,
-  });
-});
+userRoutes.post("/", adaptRoute(makeAddUserController()));
